+++ web/src/common/enums.ts
@@ -1,0 +1,6 @@
+export enum PFSize {
+    Small = "pf-m-sm",
+    Medium = "pf-m-md",
+    Large = "pf-m-lg",
+    XLarge = "pf-m-xl",
+}
