+++ web/src/elements/Spinner.ts
@@ -1,3 +1,4 @@
+import { PFSize } from "@goauthentik/common/enums.js";
 import { AKElement } from "@goauthentik/elements/Base";
 
 import { msg } from "@lit/localize";
@@ -5,13 +6,6 @@ import { CSSResult, TemplateResult, html } from "lit";
 import { customElement, property } from "lit/decorators.js";
 
 import PFSpinner from "@patternfly/patternfly/components/Spinner/spinner.css";
-
-export enum PFSize {
-    Small = "pf-m-sm",
-    Medium = "pf-m-md",
-    Large = "pf-m-lg",
-    XLarge = "pf-m-xl",
-}
 
 @customElement("ak-spinner")
 export class Spinner extends AKElement {
