+++ web/src/elements/buttons/SpinnerButton/BaseTaskButton.ts
@@ -1,6 +1,6 @@
 import { ERROR_CLASS, PROGRESS_CLASS, SUCCESS_CLASS } from "@goauthentik/common/constants";
+import { PFSize } from "@goauthentik/common/enums.js";
 import { AKElement } from "@goauthentik/elements/Base";
-import { PFSize } from "@goauthentik/elements/Spinner";
 import { CustomEmitterElement } from "@goauthentik/elements/utils/eventEmitter";
 
 import { Task, TaskStatus } from "@lit/task";
